==> 2 - dados PNAD IBGE.ipynb <==
import pandas as pd
import requests
import numpy as np

url1 = 'https://apisidra.ibge.gov.br/values/t/4092/n3/all/v/1641/p/202001,202002,202003,202004/c629/32386,32446'
url2 = 'https://apisidra.ibge.gov.br/values/t/4097/n3/all/v/4090/p/202001,202002,202003,202004/c11913/31721,31724,31727,31731,96170,96171'

Tabela 4092

resp = requests.get(url1).json()

data = pd.DataFrame(resp)

data = data.rename(columns={"D1N": "uf_nome", "V": "valor", "D3C": "trimestre", "D4N": "condicao", "D2N": "variavel"})

data = data[["trimestre", "uf_nome", "variavel", "condicao", "valor"]].iloc[1:]

conditions = [
    (data['uf_nome'] == 'Acre'), (data['uf_nome'] == 'Amazonas'), (data['uf_nome'] == 'Roraima'), (data['uf_nome'] == 'Pará'), (data['uf_nome'] == 'Amapá'), (data['uf_nome'] == 'Tocantins'), 
    (data['uf_nome'] == 'Maranhão'), (data['uf_nome'] == 'Piauí'), (data['uf_nome'] == 'Ceará'), (data['uf_nome'] == 'Rio Grande do Norte'), (data['uf_nome'] == 'Paraíba'), (data['uf_nome'] == 'Pernambuco'), 
    (data['uf_nome'] == 'Alagoas'), (data['uf_nome'] == 'Sergipe'), (data['uf_nome'] == 'Bahia'), (data['uf_nome'] == 'Minas Gerais'), (data['uf_nome'] == 'Espírito Santo'), 
    (data['uf_nome'] == 'Rio de Janeiro'), (data['uf_nome'] == 'São Paulo'), (data['uf_nome'] == 'Paraná'), (data['uf_nome'] == 'Santa Catarina'), (data['uf_nome'] == 'Rondônia'),
    (data['uf_nome'] == 'Rio Grande do Sul'), (data['uf_nome'] == 'Mato Grosso do Sul'), (data['uf_nome'] == 'Mato Grosso'), (data['uf_nome'] == 'Goiás'), (data['uf_nome'] == 'Distrito Federal')]

values = ['AC', 'AM', 'RR', 'PA', 'AP', 'TO', 'MA', 'PI', 'CE', 'RN', 'PB', 'PE', 'AL', 'SE', 'BA', 'MG', 'ES', 'RJ', 'SP', 'PR', 'SC', 'RO', 'RS', 'MS', 'MT', 'GO', 'DF']

data['uf'] = np.select(conditions, values)

conditions = [
        (data['trimestre'] == '202001'), (data['trimestre'] == '202002'), (data['trimestre'] == '202003'), (data['trimestre'] == '202004')]

values = ['202003', '202006', '202009',' 202012']

data['trimestre_fim'] = np.select(conditions, values)

data.to_csv('data_pnad1.csv')

Tabela 4097

resp = requests.get(url2).json()

data2 = pd.DataFrame(resp)

data2 = data2.rename(columns={"D1N": "uf_nome", "V": "valor", "D3C": "trimestre", "D4N": "condicao", "D2N": "variavel"})

data2 = data2[["trimestre", "uf_nome", "variavel", "condicao", "valor"]].iloc[1:]

conditions = [
    (data2['uf_nome'] == 'Acre'), (data2['uf_nome'] == 'Amazonas'), (data2['uf_nome'] == 'Roraima'), (data2['uf_nome'] == 'Pará'), (data2['uf_nome'] == 'Amapá'), (data2['uf_nome'] == 'Tocantins'), 
    (data2['uf_nome'] == 'Maranhão'), (data2['uf_nome'] == 'Piauí'), (data2['uf_nome'] == 'Ceará'), (data2['uf_nome'] == 'Rio Grande do Norte'), (data2['uf_nome'] == 'Paraíba'), 
    (data2['uf_nome'] == 'Pernambuco'), (data2['uf_nome'] == 'Alagoas'), (data2['uf_nome'] == 'Sergipe'), (data2['uf_nome'] == 'Bahia'), (data2['uf_nome'] == 'Minas Gerais'), 
    (data2['uf_nome'] == 'Espírito Santo'), (data2['uf_nome'] == 'Rio de Janeiro'), (data2['uf_nome'] == 'São Paulo'), (data2['uf_nome'] == 'Paraná'), (data2['uf_nome'] == 'Santa Catarina'), 
    (data2['uf_nome'] == 'Rondônia'), (data2['uf_nome'] == 'Rio Grande do Sul'), (data2['uf_nome'] == 'Mato Grosso do Sul'), (data2['uf_nome'] == 'Mato Grosso'), (data2['uf_nome'] == 'Goiás'), 
    (data2['uf_nome'] == 'Distrito Federal')]

values = ['AC', 'AM', 'RR', 'PA', 'AP', 'TO', 'MA', 'PI', 'CE', 'RN', 'PB', 'PE', 'AL', 'SE', 'BA', 'MG', 'ES', 'RJ', 'SP', 'PR', 'SC', 'RO', 'RS', 'MS', 'MT', 'GO', 'DF']

data2['uf'] = np.select(conditions, values)

conditions = [
        (data2['trimestre'] == '202001'), (data2['trimestre'] == '202002'), (data2['trimestre'] == '202003'), (data2['trimestre'] == '202004')]

values = ['202003', '202006', '202009',' 202012']

data2['trimestre_fim'] = np.select(conditions, values)

data2.to_csv('data_pnad2.csv')
